% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{.assert_surv_design}
\alias{.assert_surv_design}
\title{Assert object inherits from surv_design}
\usage{
.assert_surv_design(x, arg_name = "design")
}
\arguments{
\item{x}{Object to check.}

\item{arg_name}{Name for error messages.}
}
\value{
Invisible TRUE.
}
\description{
Assert object inherits from surv_design
}
\keyword{internal}
