% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{.date_to_epiweek}
\alias{.date_to_epiweek}
\title{Convert Date vector to ISO epiweek strings (YYYY-Wnn)}
\usage{
.date_to_epiweek(dates)
}
\arguments{
\item{dates}{Date vector.}
}
\value{
Character vector of epiweek labels.
}
\description{
Convert Date vector to ISO epiweek strings (YYYY-Wnn)
}
\keyword{internal}
