% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-weights.R
\name{.map_weights_to_obs}
\alias{.map_weights_to_obs}
\title{Map stratum-level weights to individual observations}
\usage{
.map_weights_to_obs(design)
}
\arguments{
\item{design}{A surv_design object.}
}
\value{
Tibble with all data columns plus stratum_id, seq_rate, weight.
}
\description{
Map stratum-level weights to individual observations
}
\keyword{internal}
