% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{.resolve_time_column}
\alias{.resolve_time_column}
\title{Resolve user-specified time column}
\usage{
.resolve_time_column(dat, time, date_col)
}
\arguments{
\item{dat}{Data frame.}

\item{time}{User specification string.}

\item{date_col}{Name of the date column in dat.}
}
\value{
List with data (possibly modified) and col (column name).
}
\description{
Handles "epiweek", "month", "date", or a literal column name.
}
\keyword{internal}
