% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-pub-theme.R
\name{.surv_colors}
\alias{.surv_colors}
\title{survinger color palette}
\usage{
.surv_colors()
}
\value{
A named character vector of hex colors.
}
\description{
survinger color palette
}
\keyword{internal}
