% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-tidy-methods.R
\name{glance.surv}
\alias{glance.surv}
\alias{glance.surv_prevalence}
\alias{glance.surv_delay_fit}
\alias{glance.surv_adjusted}
\title{One-row summary of survinger model}
\usage{
\method{glance}{surv_prevalence}(x, ...)

\method{glance}{surv_delay_fit}(x, ...)

\method{glance}{surv_adjusted}(x, ...)
}
\arguments{
\item{x}{A survinger result object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A single-row tibble with key summary statistics.
}
\description{
One-row summary of survinger model
}
\examples{
sim <- surv_simulate(n_regions = 3, n_weeks = 10, seed = 1)
d <- surv_design(sim$sequences, ~ region,
                 sim$population[c("region", "seq_rate")], sim$population)
prev <- surv_lineage_prevalence(d, "BA.2.86")
glance(prev)
}
