% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.R
\name{new_surv_adjusted}
\alias{new_surv_adjusted}
\title{Internal constructor for surv_adjusted}
\usage{
new_surv_adjusted(estimates, design_component, delay_component, method)
}
\description{
Internal constructor for surv_adjusted
}
\keyword{internal}
