% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.R
\name{new_surv_allocation}
\alias{new_surv_allocation}
\title{Internal constructor for surv_allocation}
\usage{
new_surv_allocation(
  allocation,
  objective,
  total_capacity,
  constraints,
  diagnostics
)
}
\description{
Internal constructor for surv_allocation
}
\keyword{internal}
