% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.R
\name{new_surv_delay_fit}
\alias{new_surv_delay_fit}
\title{Internal constructor for surv_delay_fit}
\usage{
new_surv_delay_fit(distribution, parameters, strata, data_summary, diagnostics)
}
\description{
Internal constructor for surv_delay_fit
}
\keyword{internal}
