% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.R
\name{new_surv_nowcast}
\alias{new_surv_nowcast}
\title{Internal constructor for surv_nowcast}
\usage{
new_surv_nowcast(estimates, delay_fit, truncation_window, method, lineage)
}
\description{
Internal constructor for surv_nowcast
}
\keyword{internal}
