% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.R
\name{new_surv_prevalence}
\alias{new_surv_prevalence}
\title{Internal constructor for surv_prevalence}
\usage{
new_surv_prevalence(estimates, design, method, lineage, conf_level, time_unit)
}
\description{
Internal constructor for surv_prevalence
}
\keyword{internal}
