% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-plot-methods.R
\name{plot.surv}
\alias{plot.surv}
\alias{plot.surv_design}
\alias{plot.surv_allocation}
\alias{plot.surv_prevalence}
\alias{plot.surv_delay_fit}
\alias{plot.surv_nowcast}
\alias{plot.surv_adjusted}
\title{Plot methods for survinger objects}
\usage{
\method{plot}{surv_design}(x, ...)

\method{plot}{surv_allocation}(x, ...)

\method{plot}{surv_prevalence}(x, ...)

\method{plot}{surv_delay_fit}(x, ...)

\method{plot}{surv_nowcast}(x, ...)

\method{plot}{surv_adjusted}(x, ...)
}
\arguments{
\item{x}{A survinger object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A ggplot2 object.
}
\description{
Visualize surveillance design, allocation, prevalence estimates,
delay distributions, nowcasts, and adjusted estimates.
}
