% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-diagnostics.R
\name{surv_design_effect}
\alias{surv_design_effect}
\title{Compute design effect over time}
\usage{
surv_design_effect(weighted, naive)
}
\arguments{
\item{weighted}{A \code{surv_prevalence} object.}

\item{naive}{A \code{surv_prevalence} object.}
}
\value{
A tibble with time, deff, and bias_correction columns.
}
\description{
Compute design effect over time
}
\examples{
sim <- surv_simulate(n_regions = 3, n_weeks = 10, seed = 1)
d <- surv_design(sim$sequences, ~ region,
                 sim$population[c("region", "seq_rate")], sim$population)
w <- surv_lineage_prevalence(d, "BA.2.86")
n <- surv_naive_prevalence(d, "BA.2.86")
surv_design_effect(w, n)

}
