% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-design.R
\name{surv_filter}
\alias{surv_filter}
\title{Subset a surveillance design by filter criteria}
\usage{
surv_filter(design, ...)
}
\arguments{
\item{design}{A \code{surv_design} object.}

\item{...}{Filter expressions passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}
}
\value{
A new \code{surv_design} object with filtered data.
}
\description{
Creates a new \code{surv_design} object containing only sequences
matching the specified filter criteria.
}
\examples{
sim <- surv_simulate(n_regions = 5, n_weeks = 12, seed = 1)
d <- surv_design(sim$sequences, ~ region,
                 sim$population[c("region", "seq_rate")], sim$population)
d_sub <- surv_filter(d, region \%in\% c("Region_A", "Region_B"))
print(d_sub)

}
\seealso{
\code{\link[=surv_design]{surv_design()}}
}
