% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-diagnostics.R
\name{surv_plot_allocation}
\alias{surv_plot_allocation}
\title{Plot allocation plan}
\usage{
surv_plot_allocation(allocation)
}
\arguments{
\item{allocation}{A \code{surv_allocation} object.}
}
\value{
A ggplot2 object.
}
\description{
Plot allocation plan
}
\examples{
sim <- surv_simulate(n_regions = 5, n_weeks = 10, seed = 1)
d <- surv_design(sim$sequences, ~ region,
                 sim$population[c("region", "seq_rate")], sim$population)
a <- surv_optimize_allocation(d, "min_mse", total_capacity = 500)
surv_plot_allocation(a)

}
