% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-detection.R
\name{surv_required_sequences}
\alias{surv_required_sequences}
\title{Required sequences for target detection probability}
\usage{
surv_required_sequences(
  true_prevalence,
  target_probability = 0.95,
  n_periods = 1L,
  detection_threshold = 1L
)
}
\arguments{
\item{true_prevalence}{Numeric.}

\item{target_probability}{Numeric. Default 0.95.}

\item{n_periods}{Integer. Default 1.}

\item{detection_threshold}{Integer. Default 1.}
}
\value{
Integer.
}
\description{
Required sequences for target detection probability
}
\examples{
surv_required_sequences(0.01)
surv_required_sequences(0.05, target_probability = 0.99)

}
