% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-design.R
\name{surv_set_weights}
\alias{surv_set_weights}
\title{Override design weights with custom values}
\usage{
surv_set_weights(design, weights)
}
\arguments{
\item{design}{A \code{surv_design} object.}

\item{weights}{Numeric vector of length equal to number of strata.}
}
\value{
Updated \code{surv_design}.
}
\description{
Override design weights with custom values
}
\examples{
sim <- surv_simulate(n_regions = 3, n_weeks = 4, seed = 1)
d <- surv_design(sim$sequences, ~ region,
                 sim$population[c("region", "seq_rate")], sim$population)
d2 <- surv_set_weights(d, rep(1.0, d$n_strata))

}
