% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-design.R
\name{surv_update_rates}
\alias{surv_update_rates}
\title{Update sequencing rates in a surveillance design}
\usage{
surv_update_rates(design, new_rates)
}
\arguments{
\item{design}{A \code{surv_design} object.}

\item{new_rates}{Data frame with strata columns + \code{seq_rate},
or named numeric vector.}
}
\value{
Updated \code{surv_design}.
}
\description{
Update sequencing rates in a surveillance design
}
\examples{
sim <- surv_simulate(n_regions = 3, n_weeks = 4, seed = 1)
d <- surv_design(sim$sequences, ~ region,
                 sim$population[c("region", "seq_rate")], sim$population)
new_r <- sim$population[c("region", "seq_rate")]
new_r$seq_rate <- new_r$seq_rate * 2
d2 <- surv_update_rates(d, new_r)

}
