% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-pub-theme.R
\name{theme_survinger}
\alias{theme_survinger}
\title{Publication-quality ggplot2 theme}
\usage{
theme_survinger(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size. Default 11.}

\item{base_family}{Font family. Default \code{""}.}
}
\value{
A ggplot2 theme object.
}
\description{
A clean, high-contrast theme suitable for journal publication.
Follows Nature/Lancet figure guidelines.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_survinger()

}
\seealso{
\code{\link[=surv_compare_estimates]{surv_compare_estimates()}}
}
