% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-tidy-methods.R
\name{tidy.surv}
\alias{tidy.surv}
\alias{tidy.surv_prevalence}
\alias{tidy.surv_nowcast}
\alias{tidy.surv_adjusted}
\alias{tidy.surv_allocation}
\alias{tidy.surv_delay_fit}
\title{Extract tidy estimates from survinger objects}
\usage{
\method{tidy}{surv_prevalence}(x, ...)

\method{tidy}{surv_nowcast}(x, ...)

\method{tidy}{surv_adjusted}(x, ...)

\method{tidy}{surv_allocation}(x, ...)

\method{tidy}{surv_delay_fit}(x, ...)
}
\arguments{
\item{x}{A survinger result object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A tibble.
}
\description{
Converts survinger result objects into tidy tibbles suitable for
further analysis with dplyr, ggplot2, or other tidyverse tools.
}
\examples{
sim <- surv_simulate(n_regions = 3, n_weeks = 10, seed = 1)
d <- surv_design(sim$sequences, ~ region,
                 sim$population[c("region", "seq_rate")], sim$population)
prev <- surv_lineage_prevalence(d, "BA.2.86")
tidy(prev)
}
